use onLeft, onNothing, onJust
