+++ server/src-lib/Hasura/RQL/DDL/Metadata.hs
@@ -30,7 +30,6 @@ import qualified Data.HashSet                   as HS
 import qualified Data.List                      as L
 import qualified Data.Text                      as T
 
-import           Hasura.GraphQL.Utils
 import           Hasura.Prelude
 import           Hasura.RQL.Types
 import           Hasura.SQL.Types
